+++ src/Ds/Component/Audit/DependencyInjection/Configuration.php
@@ -19,15 +19,6 @@ class Configuration implements ConfigurationInterface
     {
         $builder = new TreeBuilder;
         $node = $builder->root('ds_audit');
-        $node
-            ->children()
-                ->arrayNode('audit')
-                    ->children()
-                        ->scalarNode('owner')->end()
-                        ->scalarNode('owner_uuid')->end()
-                    ->end()
-                ->end()
-            ->end();
 
         return $builder;
     }