+++ src/Ds/Component/Config/DependencyInjection/Configuration.php
@@ -19,6 +19,35 @@ class Configuration implements ConfigurationInterface
     {
         $builder = new TreeBuilder;
         $node = $builder->root('ds_config');
+        $node
+            ->children()
+                ->arrayNode('configs')
+                    ->prototype('array')
+                        ->children()
+                            ->scalarNode('title')
+                            ->end()
+                            ->scalarNode('key')
+                            ->end()
+                            ->booleanNode('encrypt')
+                                ->defaultFalse()
+                            ->end()
+                        ->end()
+                    ->end()
+                ->end()
+                ->arrayNode('parameters')
+                    ->prototype('array')
+                        ->children()
+                            ->scalarNode('title')
+                            ->end()
+                            ->scalarNode('key')
+                            ->end()
+                            ->booleanNode('encrypt')
+                                ->defaultFalse()
+                            ->end()
+                        ->end()
+                    ->end()
+                ->end()
+            ->end();
 
         return $builder;
     }