+++ src/Ds/Component/Config/DsConfigBundle.php
@@ -2,8 +2,6 @@
 
 namespace Ds\Component\Config;
 
-use Ds\Component\Config\DependencyInjection\Compiler;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 /**
@@ -13,13 +11,4 @@ use Symfony\Component\HttpKernel\Bundle\Bundle;
  */
 class DsConfigBundle extends Bundle
 {
-    /**
-     * {@inheritdoc}
-     */
-    public function build(ContainerBuilder $container)
-    {
-        parent::build($container);
-        $container->addCompilerPass(new Compiler\ConfigPass);
-        $container->addCompilerPass(new Compiler\ParameterPass);
-    }
 }