+++ src/Ds/Component/Encryption/DependencyInjection/DsEncryptionExtension.php
@@ -19,9 +19,6 @@ final class DsEncryptionExtension extends Extension
      */
     public function load(array $configs, ContainerBuilder $container)
     {
-        $configuration = new Configuration;
-        $config = $this->processConfiguration($configuration, $configs);
-
         $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
         $loader->load('parameters.yml');
         $loader->load('event_listeners.yml');