+++ src/Ds/Component/Encryption/Model/Type/Encryptable.php
@@ -9,4 +9,18 @@ namespace Ds\Component\Encryption\Model\Type;
  */
 interface Encryptable
 {
+    /**
+     * Set encrypted status
+     *
+     * @param boolean $encrypted
+     * @return \Ds\Component\Encryption\Model\Type\Encryptable
+     */
+    public function setEncrypted(bool $encrypted) : Encryptable;
+
+    /**
+     * Get encrypted status
+     *
+     * @return boolean
+     */
+    public function getEncrypted() : bool;
 }